fix: Wrong hash renamer file name
