+++ src/lib/hash-renamer.ts
